Create Housekeeping
This contract will be used to scan a recall the existing bets that users have currently open and closed bets. These functions should allow us to keep transaction costs low as they are not writing or processing on chain, simply reading from what is there. Obviously still need to add the function to scan closed contracts, but currently unsure of how to implement. May need to take hybrid chain approach?